+++ app-wearos/src/main/java/de/danoeh/antennapod/wearos/MainViewModel.kt
@@ -7,6 +7,7 @@ import de.danoeh.antennapod.net.sync.wearinterface.WearConnectionUtils
 import de.danoeh.antennapod.net.sync.wearinterface.WearDataPaths
 import de.danoeh.antennapod.net.sync.wearinterface.WearNowPlaying
 import de.danoeh.antennapod.wearos.sync.WearDataRepository
+import de.danoeh.antennapod.wearos.sync.WearMessageSender
 import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
@@ -22,14 +23,14 @@ class MainViewModel(application: Application) : AndroidViewModel(application) {
     val uiState: StateFlow<MainUiState> = _uiState
 
     init {
-        viewModelScope.launch(Dispatchers.Default) {
+        viewModelScope.launch(Dispatchers.IO) {
             val name = WearConnectionUtils.getConnectedNodeName(getApplication())
             _uiState.update { it.copy(connectedPhone = name) }
         }
 
         viewModelScope.launch {
             while (true) {
-                WearDataRepository.sendMessage(getApplication(), WearDataPaths.NOW_PLAYING)
+                WearMessageSender.send(getApplication(), WearDataPaths.NOW_PLAYING)
                 delay(5.seconds)
             }
         }
