+++ app-wearos/src/main/java/de/danoeh/antennapod/wearos/sync/WearDataListenerService.kt
@@ -9,18 +9,24 @@ import de.danoeh.antennapod.net.sync.wearinterface.WearSerializer
 class WearDataListenerService : WearableListenerService() {
     override fun onMessageReceived(event: MessageEvent) {
         val path = event.path
-        if (path == WearDataPaths.SUBSCRIPTIONS) {
-            val feeds = WearSerializer.feedsFromBytes(event.data)
-            Log.d(TAG, "Received $path: ${feeds.size} feeds")
-            WearDataRepository.updateFeeds(path, feeds)
-        } else if (path == WearDataPaths.NOW_PLAYING) {
-            WearDataRepository.updateNowPlaying(WearSerializer.nowPlayingFromBytes(event.data))
-        } else if (path == WearDataPaths.QUEUE || path == WearDataPaths.DOWNLOADS ||
-            path == WearDataPaths.EPISODES || path.startsWith(WearDataPaths.FEED_EPISODES_PREFIX)
-        ) {
-            val items = WearSerializer.episodesFromBytes(event.data)
-            Log.d(TAG, "Received $path: ${items.size} episodes")
-            WearDataRepository.updateEpisodes(path, items)
+        when (path) {
+            WearDataPaths.SUBSCRIPTIONS -> {
+                val feeds = WearSerializer.feedsFromBytes(event.data)
+                Log.d(TAG, "Received $path: ${feeds.size} feeds")
+                WearDataRepository.updateFeeds(path, feeds)
+            }
+            WearDataPaths.NOW_PLAYING ->
+                WearDataRepository.updateNowPlaying(WearSerializer.nowPlayingFromBytes(event.data))
+            WearDataPaths.QUEUE, WearDataPaths.DOWNLOADS, WearDataPaths.EPISODES -> {
+                val items = WearSerializer.episodesFromBytes(event.data)
+                Log.d(TAG, "Received $path: ${items.size} episodes")
+                WearDataRepository.updateEpisodes(path, items)
+            }
+            else -> if (path.startsWith(WearDataPaths.FEED_EPISODES_PREFIX)) {
+                val items = WearSerializer.episodesFromBytes(event.data)
+                Log.d(TAG, "Received $path: ${items.size} episodes")
+                WearDataRepository.updateEpisodes(path, items)
+            }
         }
     }
