+++ app-wearos/src/main/java/de/danoeh/antennapod/wearos/sync/WearDataRepository.kt
@@ -1,14 +1,10 @@
 package de.danoeh.antennapod.wearos.sync
 
-import android.content.Context
-import android.util.Log
-import com.google.android.gms.wearable.Wearable
 import de.danoeh.antennapod.model.feed.Feed
 import de.danoeh.antennapod.model.feed.FeedItem
 import de.danoeh.antennapod.net.sync.wearinterface.WearNowPlaying
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.tasks.await
 
 /**
  * Singleton holding the last data received from the phone for each data path.
@@ -16,8 +12,6 @@ import kotlinx.coroutines.tasks.await
  * read by the screen composables.
  */
 object WearDataRepository {
-    private const val TAG = "WearDataRepository"
-
     private val _episodesByPath = MutableStateFlow<Map<String, List<FeedItem>>>(emptyMap())
     val episodesByPath: StateFlow<Map<String, List<FeedItem>>> = _episodesByPath
 
@@ -37,18 +31,5 @@ object WearDataRepository {
 
     fun updateFeeds(path: String, feeds: List<Feed>) {
         _feedsByPath.value += (path to feeds)
-    }
-
-    suspend fun sendMessage(context: Context, path: String) {
-        try {
-            val nodes = Wearable.getNodeClient(context).connectedNodes.await()
-            val node = nodes.firstOrNull() ?: run {
-                Log.w(TAG, "No connected nodes")
-                return
-            }
-            Wearable.getMessageClient(context).sendMessage(node.id, path, null).await()
-        } catch (e: Exception) {
-            Log.e(TAG, "Failed to send message to phone", e)
-        }
     }
 }
